added dispatchToHandlerWith, matchWith, resourceHandlerInstance as abstract methods to Resource class moved allocateHandlerPool, pooledHandler to Resource class
