+++ src/main/java/io/vlingo/http/resource/Resource.java
@@ -1,7 +1,51 @@
 package io.vlingo.http.resource;
 
+import io.vlingo.actors.Definition;
+import io.vlingo.actors.Stage;
 import io.vlingo.http.Context;
+import io.vlingo.http.Method;
+
+import java.net.URI;
+import java.util.concurrent.atomic.AtomicLong;
 
 public abstract class Resource<T> {
-    public abstract void dispatchToHandlerWith(final Context context, final Action.MappedParameters mappedParameters);
+  public final String name;
+  public final Class<? extends ResourceHandler> resourceHandlerClass;
+  public final int handlerPoolSize;
+
+  private final ResourceRequestHandler[] handlerPool;
+  private final AtomicLong handlerPoolIndex;
+
+  public abstract void dispatchToHandlerWith(final Context context, final Action.MappedParameters mappedParameters);
+
+  abstract Action.MatchResults matchWith(final Method method, final URI uri);
+
+  protected abstract ResourceHandler resourceHandlerInstance(final Stage stage);
+
+  void allocateHandlerPool(final Stage stage) {
+    for (int idx = 0; idx < handlerPoolSize; ++idx) {
+      handlerPool[idx] =
+        stage.actorFor(
+          Definition.has(
+            ResourceRequestHandlerActor.class,
+            Definition.parameters(resourceHandlerInstance(stage))),
+          ResourceRequestHandler.class);
+    }
+  }
+
+  protected ResourceRequestHandler pooledHandler() {
+    final int index = (int) (handlerPoolIndex.incrementAndGet() % handlerPoolSize);
+    return handlerPool[index];
+  }
+
+  Resource(final String name,
+           final Class<? extends ResourceHandler> resourceHandlerClass,
+           final int handlerPoolSize) {
+    this.name = name;
+    this.resourceHandlerClass = resourceHandlerClass;
+    this.handlerPoolSize = handlerPoolSize;
+    this.handlerPool = new ResourceRequestHandler[handlerPoolSize];
+    this.handlerPoolIndex = new AtomicLong(0);
+  }
+
 }
